requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pytorch-lightning>=2.0.0
wandb>=0.15.0
hydra-core>=1.3.0
omegaconf>=2.3.0
torchmetrics>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pillow>=10.0.0
timm>=0.9.0

